Reuse text stacktrace parsing for faster parsing of last stacktrace
The new :LocStackLast command queries the stacktrace (via nREPL)
directly from *e, which is then converted to a string, in the same way
as stacktrace data is read from a register and serialized. The
serialized string is then parsed in Fennel, as per :LocStackReg.

The biggest difference between :LocStack and :LocStackLast is therefore
that the latter doesn't do the extremely slow `stacktrace` nREPL op that
the former does, and is therefore significantly faster. The cost for
this speedup is a decrease in stacktrace data resolution. For example,
:LocStack will returns these lines:

    clojure.core|3215| clojure.core$eval/invokeStatic
    clojure.core|3211| clojure.core$eval/invoke

whereas :LocStackLast will only return these:

    clojure.core|3211| eval (clojure.core$eval)
    clojure.core|3211| eval (clojure.core$eval)

Notice how the former has the more specific line 3215.